pyproject.toml
[project]
name = "airbnb-agent"
version = "1.6.5"
description = "Agente visual de calendario Airbnb"
requires-python = ">=3.11"
authors = [{name = "edgsanchez"}]
readme = "README.md"
dependencies = [
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "icalendar>=5.0.0",
    "requests>=2.31.0",
    "pymongo>=4.0.0",
    "gunicorn>=23.0.0",
]

[tool.poetry]
packages = [{include = "airbnb_agent"}]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
